+++ crates/report/src/state.rs
@@ -49,7 +49,7 @@ pub struct State {
 
     /// Authorization pools.
     #[json(Vec<Vec<String>>)]
-    pub auth_pools: [Vec<OpaqueHash>; CORES_COUNT as usize],
+    pub auth_pools: [Vec<OpaqueHash>; CORES_COUNT],
 
     /// [δ] Encoded services dictionary. Refer to T(σ) in Appendix D.
     #[json(nested)]
